ij1.04q04

+++ src/ij/plugin/Selection.java
@@ -27,15 +27,23 @@ public class Selection implements PlugIn, Measurements {
 
 	public void run(String arg) {
 		imp = WindowManager.getCurrentImage();
-		if (arg.equals("add"))
-			{addToRoiManager(imp); return;}
-		if (imp==null)
-			{IJ.noImage(); return;}
-		if (arg.equals("all"))
-			imp.setRoi(0,0,imp.getWidth(),imp.getHeight());
-		else if (arg.equals("none"))
-			imp.deleteRoi();
-		else if (arg.equals("restore"))
+		if (arg.equals("add")) {
+			addToRoiManager(imp);
+			return;
+		}
+		if (imp==null) {
+			IJ.noImage();
+			return;
+		}
+		if (arg.equals("all")) {
+			if (imp.okToDeleteRoi()) {
+				imp.saveRoi();
+				imp.setRoi(0,0,imp.getWidth(),imp.getHeight());
+			}
+		} else if (arg.equals("none")) {
+			if (imp.okToDeleteRoi())
+				imp.deleteRoi();
+		} else if (arg.equals("restore"))
 			imp.restoreRoi();
 		else if (arg.equals("spline"))
 			fitSpline();
@@ -70,6 +78,7 @@ public class Selection implements PlugIn, Measurements {
 	}
 	
 	private void rotate(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		Roi roi = imp.getRoi();
 		if (IJ.macroRunning()) {
 			String options = Macro.getOptions();
@@ -82,6 +91,7 @@ public class Selection implements PlugIn, Measurements {
 	}
 	
 	private void enlarge(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		Roi roi = imp.getRoi();
 		if (roi!=null) {
 			Undo.setup(Undo.ROI, imp);
@@ -101,6 +111,7 @@ public class Selection implements PlugIn, Measurements {
 	Authors: Nikolai Chernov, Michael Doube, Ved Sharma
 	*/
 	void fitCircle(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		Roi roi = imp.getRoi();
 		if (roi==null) {
 			noRoi("Fit Circle");
@@ -454,6 +465,7 @@ public class Selection implements PlugIn, Measurements {
 	}
 
 	void createEllipse(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		IJ.showStatus("Fitting ellipse");
 		Roi roi = imp.getRoi();
 		if (roi==null)
@@ -479,28 +491,23 @@ public class Selection implements PlugIn, Measurements {
 	}
 
 	void convexHull(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
+		long startTime = System.currentTimeMillis();
 		Roi roi = imp.getRoi();
-		int type = roi!=null?roi.getType():-1;
-		if (!(type==Roi.FREEROI||type==Roi.TRACED_ROI||type==Roi.POLYGON||type==Roi.POINT))
-			{IJ.error("Convex Hull", "Polygonal or point selection required"); return;}
-		if (roi instanceof EllipseRoi)
-			return;
-		//if (roi.subPixelResolution() && roi instanceof PolygonRoi) {
-		//	FloatPolygon p = ((PolygonRoi)roi).getFloatConvexHull();
-		//	if (p!=null)
-		//		imp.setRoi(new PolygonRoi(p.xpoints, p.ypoints, p.npoints, roi.POLYGON));
-		//} else {
-		Polygon p = roi.getConvexHull();
+		if (roi == null) { IJ.error("Convex Hull", "Selection required"); return; }
+		if (roi instanceof Line) { IJ.error("Convex Hull", "Area selection, point selection, or segmented or free line required"); return; }
+		FloatPolygon p = roi.getFloatConvexHull();
 		if (p!=null) {
 			Undo.setup(Undo.ROI, imp);
-			Roi roi2 = new PolygonRoi(p.xpoints, p.ypoints, p.npoints, roi.POLYGON);
+			Roi roi2 = new PolygonRoi(p, roi.POLYGON);
 			transferProperties(roi, roi2);
 			imp.setRoi(roi2);
+			IJ.showTime(imp, startTime, "Convex Hull ", 1);
 		}
 	}
 	
 	// Finds the index of the upper right point that is guaranteed to be on convex hull
-	int findFirstPoint(int[] xCoordinates, int[] yCoordinates, int n, ImagePlus imp) {
+	/*int findFirstPoint(int[] xCoordinates, int[] yCoordinates, int n, ImagePlus imp) {
 		int smallestY = imp.getHeight();
 		int x, y;
 		for (int i=0; i<n; i++) {
@@ -519,7 +526,7 @@ public class Selection implements PlugIn, Measurements {
 			}
 		}
 		return p1;
-	}
+	}*/
 	
 	void createMask(ImagePlus imp) {
 		Roi roi = imp.getRoi();
@@ -540,6 +547,7 @@ public class Selection implements PlugIn, Measurements {
 		ByteProcessor mask = imp.createRoiMask();
 		if (!Prefs.blackBackground)
 			mask.invertLut();
+		mask.setThreshold(255, 255, ImageProcessor.NO_LUT_UPDATE);
 		ImagePlus maskImp = null;
 		Frame frame = WindowManager.getFrame("Mask");
 		if (frame!=null && (frame instanceof ImageWindow))
@@ -573,7 +581,16 @@ public class Selection implements PlugIn, Measurements {
 		ByteProcessor mask = imp.createThresholdMask();
 		if (!Prefs.blackBackground)
 			mask.invertLut();
-		new ImagePlus("mask",mask).show();
+		mask.setThreshold(255, 255, ImageProcessor.NO_LUT_UPDATE);
+		ImagePlus maskImp = new ImagePlus("mask",mask);
+		Calibration cal = imp.getCalibration();
+		if (cal.scaled()) {
+			Calibration cal2 = maskImp.getCalibration();
+			cal2.pixelWidth = cal.pixelWidth;
+			cal2.pixelHeight = cal.pixelHeight;
+			cal2.setUnit(cal.getUnit());
+		}
+		maskImp.show();
 		Recorder.recordCall("mask = imp.createThresholdMask();");
 	}
 
@@ -593,24 +610,19 @@ public class Selection implements PlugIn, Measurements {
 			return;
 		}
 		int threshold = ip.isInvertedLut()?255:0;
-		if (Prefs.blackBackground)
-			threshold = ip.isInvertedLut()?0:255;		
 		ip.setThreshold(threshold, threshold, ImageProcessor.NO_LUT_UPDATE);
 		IJ.runPlugIn("ij.plugin.filter.ThresholdToSelection", "");
 	}
-
+	
 	void invert(ImagePlus imp) {
 		Roi roi = imp.getRoi();
-		if (roi==null || !roi.isArea())
+		if (roi == null)
+			{run("all"); return;}
+		if (!roi.isArea())
 			{IJ.error("Inverse", "Area selection required"); return;}
-		ShapeRoi s1, s2;
-		if (roi instanceof ShapeRoi)
-			s1 = (ShapeRoi)roi;
-		else
-			s1 = new ShapeRoi(roi);
-		s2 = new ShapeRoi(new Roi(0,0, imp.getWidth(), imp.getHeight()));
+		Roi inverse = roi.getInverse(imp);
 		Undo.setup(Undo.ROI, imp);
-		imp.setRoi(s1.xor(s2));
+		imp.setRoi(inverse);
 	}
 	
 	private void lineToArea(ImagePlus imp) {
@@ -693,6 +705,8 @@ public class Selection implements PlugIn, Measurements {
 			noRoi("To Bounding Box");
 			return;
 		}
+		if (!imp.okToDeleteRoi())
+			return;
 		Undo.setup(Undo.ROI, imp);
 		Rectangle r = roi.getBounds();
 		imp.deleteRoi();
